pyproject.toml
[project]
name = "flowpilot"
version = "0.1.0"
description = "AI Agent for programmer workflow automation"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
keywords = ["ai", "agent", "automation", "devops", "ssh"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # LLM SDKs
    "anthropic>=0.40.0",              # Claude
    "google-genai>=1.0.0",            # Gemini (新版 SDK)
    "zhipuai>=2.1.0",                 # 智谱 GLM
    # 核心框架
    "pydantic>=2.8",                  # Schema 校验与配置
    "pydantic-settings>=2.4",         # 环境变量配置
    # CLI
    "typer>=0.12.0",                  # CLI 框架
    "rich>=13.7",                     # 美化输出
    "prompt-toolkit>=3.0",            # 交互式输入
    # 数据存储
    "sqlalchemy>=2.0",                # ORM
    "alembic>=1.13",                  # 数据库迁移
    # SSH & 工具
    "paramiko>=3.4",                  # SSH 客户端
    "asyncssh>=2.14",                 # 异步 SSH（可选）
    # 配置解析
    "pyyaml>=6.0",                    # YAML 解析
    "python-dotenv>=1.0",             # .env 文件支持
    # 实用工具
    "httpx>=0.27",                    # HTTP 客户端（K8s API）
    "tenacity>=8.2",                  # 重试机制
    "python-dateutil>=2.8",           # 时间解析
    "zai-sdk>=0.0.3.4",
    "google-genai>=1.57.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "pytest-asyncio>=0.23",
    "ruff>=0.6.0",                    # Linter & Formatter
    "mypy>=1.11",                     # 类型检查
    "ipython>=8.20",                  # REPL
]

k8s = [
    "kubernetes>=29.0",               # K8s Python Client
]

[project.scripts]
flowpilot = "flowpilot.cli.main:app"
agent = "flowpilot.cli.main:app"      # 别名

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/flowpilot"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.6.0",
    "mypy>=1.11",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=src/flowpilot",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
]
